pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "artifactforge"
version = "0.1.0"
description = "Unified knowledge artifact generation pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ArtifactForge Team"}
]

dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "structlog>=24.0",
    "matplotlib>=3.8",
    "mermaid-py>=0.11",
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "ruff>=0.3",
    "mypy>=1.8",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
artifactforge = "artifactforge.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["artifactforge"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.mypy]
python-version = "3.11"
strict = true

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end integration tests",
    "slow: tests that take longer to run",
]
